Вынес рендер точек и фильтров в отдельные модули
